Move purger API endpoints to compactor
