+++ pkg/compactor/tenant_deletion_api.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package purger
+package compactor
 
 import (
 	"context"
@@ -11,12 +11,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/thanos-io/thanos/pkg/objstore"
 
 	"github.com/grafana/dskit/tenant"
 
@@ -25,30 +22,7 @@ import (
 	"github.com/grafana/mimir/pkg/util"
 )
 
-type TenantDeletionAPI struct {
-	bucketClient objstore.Bucket
-	logger       log.Logger
-	cfgProvider  bucket.TenantConfigProvider
-}
-
-func NewTenantDeletionAPI(storageCfg mimir_tsdb.BlocksStorageConfig, cfgProvider bucket.TenantConfigProvider, logger log.Logger, reg prometheus.Registerer) (*TenantDeletionAPI, error) {
-	bucketClient, err := createBucketClient(storageCfg, logger, reg)
-	if err != nil {
-		return nil, err
-	}
-
-	return newTenantDeletionAPI(bucketClient, cfgProvider, logger), nil
-}
-
-func newTenantDeletionAPI(bkt objstore.Bucket, cfgProvider bucket.TenantConfigProvider, logger log.Logger) *TenantDeletionAPI {
-	return &TenantDeletionAPI{
-		bucketClient: bkt,
-		cfgProvider:  cfgProvider,
-		logger:       logger,
-	}
-}
-
-func (api *TenantDeletionAPI) DeleteTenant(w http.ResponseWriter, r *http.Request) {
+func (c *MultitenantCompactor) DeleteTenant(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	userID, err := tenant.TenantID(ctx)
 	if err != nil {
@@ -58,15 +32,15 @@ func (api *TenantDeletionAPI) DeleteTenant(w http.ResponseWriter, r *http.Reques
 		return
 	}
 
-	err = mimir_tsdb.WriteTenantDeletionMark(r.Context(), api.bucketClient, userID, api.cfgProvider, mimir_tsdb.NewTenantDeletionMark(time.Now()))
+	err = mimir_tsdb.WriteTenantDeletionMark(r.Context(), c.bucketClient, userID, c.cfgProvider, mimir_tsdb.NewTenantDeletionMark(time.Now()))
 	if err != nil {
-		level.Error(api.logger).Log("msg", "failed to write tenant deletion mark", "user", userID, "err", err)
+		level.Error(c.logger).Log("msg", "failed to write tenant deletion mark", "user", userID, "err", err)
 
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	level.Info(api.logger).Log("msg", "tenant deletion mark in blocks storage created", "user", userID)
+	level.Info(c.logger).Log("msg", "tenant deletion mark in blocks storage created", "user", userID)
 
 	w.WriteHeader(http.StatusOK)
 }
@@ -76,7 +50,7 @@ type DeleteTenantStatusResponse struct {
 	BlocksDeleted bool   `json:"blocks_deleted"`
 }
 
-func (api *TenantDeletionAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.Request) {
+func (c *MultitenantCompactor) DeleteTenantStatus(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	userID, err := tenant.TenantID(ctx)
 	if err != nil {
@@ -86,7 +60,7 @@ func (api *TenantDeletionAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.
 
 	result := DeleteTenantStatusResponse{}
 	result.TenantID = userID
-	result.BlocksDeleted, err = api.isBlocksForUserDeleted(ctx, userID)
+	result.BlocksDeleted, err = c.isBlocksForUserDeleted(ctx, userID)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -95,10 +69,10 @@ func (api *TenantDeletionAPI) DeleteTenantStatus(w http.ResponseWriter, r *http.
 	util.WriteJSONResponse(w, result)
 }
 
-func (api *TenantDeletionAPI) isBlocksForUserDeleted(ctx context.Context, userID string) (bool, error) {
+func (c *MultitenantCompactor) isBlocksForUserDeleted(ctx context.Context, userID string) (bool, error) {
 	var errBlockFound = errors.New("block found")
 
-	userBucket := bucket.NewUserBucketClient(userID, api.bucketClient, api.cfgProvider)
+	userBucket := bucket.NewUserBucketClient(userID, c.bucketClient, c.cfgProvider)
 	err := userBucket.Iter(ctx, "", func(s string) error {
 		s = strings.TrimSuffix(s, "/")
 
@@ -122,13 +96,4 @@ func (api *TenantDeletionAPI) isBlocksForUserDeleted(ctx context.Context, userID
 
 	// No blocks found, all good.
 	return true, nil
-}
-
-func createBucketClient(cfg mimir_tsdb.BlocksStorageConfig, logger log.Logger, reg prometheus.Registerer) (objstore.Bucket, error) {
-	bucketClient, err := bucket.NewClient(context.Background(), cfg.Bucket, "purger", logger, reg)
-	if err != nil {
-		return nil, errors.Wrap(err, "create bucket client")
-	}
-
-	return bucketClient, nil
 }
